analyzer: Move `SpdxDocumentCache` and `SpdxResolvedDocument` to utils
These are utility classes, not the package manager implementation itself.
