+++ analyzer/src/main/kotlin/managers/utils/SpdxDocumentCache.kt
@@ -17,7 +17,7 @@
  * License-Filename: LICENSE
  */
 
-package org.ossreviewtoolkit.analyzer.managers
+package org.ossreviewtoolkit.analyzer.managers.utils
 
 import java.io.File
