+++ analyzer/src/main/kotlin/managers/utils/SpdxResolvedDocument.kt
@@ -19,7 +19,7 @@
 
 @file:Suppress("TooManyFunctions")
 
-package org.ossreviewtoolkit.analyzer.managers
+package org.ossreviewtoolkit.analyzer.managers.utils
 
 import java.io.File
 import java.net.Authenticator
